Profile Management and Data Collection
I have added the functionality for Profile management and data collection. I have made my own files for this so the userProfile.h is a redundant file that is not needed, but i am still keeping it, just in case. The actual profile mangement code can be found in profile.cpp and radoTechSimulator.cpp. I am using the radoTechSimulator files as our "Control System", so tht is gonna be our main file when it comes initialiting functionalities